% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{spinyreg-class}
\alias{fitted,spinyreg-method}
\alias{predict,spinyreg-method}
\alias{print,spinyreg-method}
\alias{residuals,spinyreg-method}
\alias{show,spinyreg-method}
\alias{spinyreg-class}
\title{Class "spinyreg"}
\description{
Class of object returned by the \pkg{spinyreg} function.
}
\section{Slots}{
 \describe{

\item{\code{coefficients}:}{numeric vector of coefficients with
respect to the original input. Contains the intercept if the model
owns any.}

\item{\code{alpha}:}{numeric scalar.}

\item{\code{gamma}:}{numeric scalar.}

\item{\code{normx}:}{Vector (class \code{"numeric"}) containing the
square root of the sum of squares of each column of the design
matrix.}

\item{\code{residuals}:}{Vector of residuals.}

\item{\code{r.squared}:}{scalar giving the
coefficient of determination.}

\item{\code{fitted}:}{Vector of fitted values.}

\item{\code{monitoring}:}{List (class \code{"list"}) which
contains various indicators dealing with the optimization
process.}

\item{\code{intercept}:}{Logical which indicates if a intercept is
included in the model.}}
}

\section{Methods}{

This class comes with the usual \code{predict(object, newx, ...)},
\code{fitted(object, ...)}, \code{residuals(object, ...)},
\code{print(object, ...)} and \code{show(object)} generic (undocumented) methods.
}
\keyword{class}

