% Generated by roxygen2 (4.0.2): do not edit by hand
\name{spinyReg}
\alias{spinyReg}
\alias{spinyreg}
\title{spinyReg}
\usage{
spinyreg(X, Y, alpha = 0.1, gamma = 1, z = rep(1, ncol(X)),
  intercept = TRUE, normalize = TRUE, verbose = 1, maxit = 1000,
  eps = 1e-12)
}
\arguments{
\item{X}{matrix of features. Do NOT include intercept.}

\item{Y}{matrix of responses.}

\item{alpha}{numeric}

\item{gamma}{numeric}

\item{z}{vector}

\item{intercept}{logical; indicates if a vector of intercepts
should be included in the model. Default is \code{TRUE}.}

\item{normalize}{logical; indicates if predictor variables should
be normalized to have unit L2 norm before fitting.  Default is
\code{TRUE}.}

\item{verbose}{integer; activate verbose mode from '0' (nothing)
to '2' (detailed output).}

\item{maxit}{integer; the maximal number of iteration
(i.e. number of alternated optimization between each parameter)
in the Expectation/Maximization algorithm.}

\item{eps}{a threshold for convergence. Default is \code{1e-12}.}
}
\value{
an object with class \code{spinyreg}, see the
documentation page \code{\linkS4class{spinyreg}} for details.
}
\description{
Computethe path of solution of a spinyReg fit.
}
\examples{
\dontrun{
data <- read.table(file="http://statweb.stanford.edu/~tibs/ElemStatLearn/datasets/prostate.data")
x <- data[, 1:8]
y <- data[, 9]
out <- spinyreg(x,y,verbose=2)
}
}
\seealso{
See also \code{\linkS4class{spinyreg}}.
}
\keyword{models,}
\keyword{regression}

